@@ -17,8 +17,8 @@
     "@material-ui/core": "4.11.3",
     "@opencollective/taxes": "3.1.0",
     "@popperjs/core": "2.6.0",
-    "@sentry/browser": "6.0.2",
-    "@sentry/node": "6.0.2",
+    "@sentry/browser": "6.0.4",
+    "@sentry/node": "6.0.4",
     "@styled-icons/boxicons-regular": "10.23.0",
     "@styled-icons/fa-brands": "10.26.0",
     "@styled-icons/fa-solid": "10.26.0",
@@ -65,7 +65,7 @@
     "lodash": "4.17.20",
     "memoize-one": "5.1.1",
     "moment": "2.29.1",
-    "next": "10.0.5",
+    "next": "10.0.6",
     "next-routes": "1.4.2",
     "node-fetch": "2.6.1",
     "nprogress": "0.2.0",
@@ -81,8 +81,8 @@
     "react-dropzone": "11.2.4",
     "react-easy-crop": "3.3.1",
     "react-flip-move": "3.0.4",
-    "react-geosuggest": "2.14.0",
-    "react-intl": "5.10.19",
+    "react-geosuggest": "2.14.1",
+    "react-intl": "5.12.0",
     "react-popper": "2.2.4",
     "react-quill": "1.3.5",
     "react-scrollchor": "6.0.0",
@@ -173,7 +173,7 @@
     "cz-conventional-changelog": "^3.3.0",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.6",
-    "eslint": "^7.18.0",
+    "eslint": "^7.19.0",
     "eslint-config-opencollective": "^2.2.0",
     "eslint-plugin-babel": "^5.3.1",
     "eslint-plugin-cypress": "^2.11.2",
