@@ -1,38 +1,79 @@
-const {
-  getDefaultSectionsForCollective,
-  filterSectionsForUser,
-  filterSectionsByData,
-  getFilteredSectionsForCollective,
-  convertSectionsToNewFormat,
-} = require('../collective-sections');
-const { CollectiveType } = require('../constants/collectives');
+import { cloneDeep, isEmpty, unset } from 'lodash';
 
-const features = {
-  COLLECTIVE_GOALS: 'DISABLED',
-  CONVERSATIONS: 'DISABLED',
-  UPDATES: 'DISABLED',
+import {
+  addDefaultSections,
+  getFilteredSectionsForCollective,
+  getSectionPath,
+  hasSection,
+} from '../collective-sections';
+import { CollectiveType } from '../constants/collectives';
+
+const expectSectionsToContain = (sections, expectedSectionsNames) => {
+  expectedSectionsNames.forEach(section => {
+    try {
+      expect(hasSection(sections, section)).toEqual(true);
+    } catch {
+      throw new Error(`Expected sections to contain "${section}", got: ${JSON.stringify(sections)}`);
+    }
+  });
+};
+
+const expectSectionsToMatch = (sections, expectedSectionsNames) => {
+  let unseenSections = cloneDeep(sections);
+  expectedSectionsNames.forEach(section => {
+    try {
+      expect(hasSection(sections, section)).toEqual(true);
+      const sectionPath = getSectionPath(sections, section);
+      unset(unseenSections, sectionPath);
+    } catch {
+      throw new Error(`Expected sections to contain "${section}", got: ${JSON.stringify(sections)}`);
+    }
+  });
+
+  // Remove empty sections & categories
+  unseenSections.forEach(e => {
+    if (e.type === 'CATEGORY') {
+      e.sections = e.sections.filter(Boolean);
+    }
+  });
+  unseenSections = unseenSections.filter(e => e.type !== 'CATEGORY' || e.sections.length > 0);
+
+  if (!isEmpty(unseenSections)) {
+    unseenSections.forEach(s => {
+      if (s.type === 'CATEGORY') {
+        console.error('Category', s.name, s.sections);
+      } else {
+        console.error(s);
+      }
+    });
+    throw new Error(`Got these extra sections`);
+  }
 };
 
 describe('Collective sections', () => {
-  describe('getDefaultSectionsForCollective', () => {
+  describe('addDefaultSections', () => {
     describe('COLLECTIVE', () => {
       it('Returns the default sections', () => {
-        expect(getDefaultSectionsForCollective(CollectiveType.COLLECTIVE, false)).toEqual([
+        expectSectionsToMatch(addDefaultSections({ type: CollectiveType.COLLECTIVE }, []), [
           'goals',
           'contribute',
+          'top-financial-contributors',
+          'events',
+          'connected-collectives',
           'updates',
           'conversations',
           'budget',
           'contributors',
           'recurring-contributions',
           'about',
+          'our-team',
         ]);
       });
     });
 
     describe('USER', () => {
       it('Also works with INDIVIDUAL', () => {
-        expect(getDefaultSectionsForCollective('INDIVIDUAL', false)).toEqual([
+        expectSectionsToMatch(addDefaultSections({ type: 'INDIVIDUAL' }, []), [
           'contributions',
           'transactions',
           'recurring-contributions',
@@ -43,252 +84,83 @@ describe('Collective sections', () => {
 
     describe('ORGANIZATION', () => {
       it('Returns the default sections', () => {
-        expect(getDefaultSectionsForCollective(CollectiveType.ORGANIZATION, false)).toEqual([
+        expectSectionsToMatch(addDefaultSections({ type: CollectiveType.ORGANIZATION }, []), [
           'contributions',
-          'contributors',
-          'updates',
-          'conversations',
-          'transactions',
           'recurring-contributions',
+          'connected-collectives',
           'about',
+          'our-team',
         ]);
       });
 
       it('Returns the default sections when active', () => {
-        expect(getDefaultSectionsForCollective(CollectiveType.ORGANIZATION, true)).toEqual([
-          'contribute',
-          'contributions',
-          'contributors',
-          'updates',
-          'conversations',
-          'transactions',
-          'budget',
-          'recurring-contributions',
-          'about',
-        ]);
+        expectSectionsToMatch(
+          addDefaultSections({ type: CollectiveType.ORGANIZATION, isHost: true, isActive: true }, []),
+          [
+            'contribute',
+            'connected-collectives',
+            'top-financial-contributors',
+            'events',
+            'updates',
+            'conversations',
+            'budget',
+            'contributions',
+            'recurring-contributions',
+            'about',
+            'our-team',
+          ],
+        );
       });
     });
 
     it('Returns an empty array for unknown types', () => {
-      expect(getDefaultSectionsForCollective()).toEqual([]);
-    });
-  });
-
-  describe('filterSectionsForUser', () => {
-    it('Works with legacy sections', () => {
-      const sections = ['contribute', 'updates', 'budget'];
-      expect(filterSectionsForUser(sections, false, false)).toEqual(sections);
-      expect(filterSectionsForUser(sections, true, false)).toEqual(sections);
-      expect(filterSectionsForUser(sections, false, true)).toEqual(sections);
-      expect(filterSectionsForUser(sections, true, true)).toEqual(sections);
-    });
-
-    it('Filters disabled sections', () => {
-      const sections = [
-        { section: 'contribute', isEnabled: true },
-        { section: 'updates', isEnabled: false },
-        { section: 'budget', isEnabled: true },
-      ];
-      const expected = ['contribute', 'budget'];
-      expect(filterSectionsForUser(sections, false, false)).toEqual(expected);
-      expect(filterSectionsForUser(sections, true, true)).toEqual(expected);
-    });
-
-    it('Filters if users do not have access', () => {
-      const sections = [
-        { section: 'contribute', isEnabled: true },
-        { section: 'updates', isEnabled: true, restrictedTo: ['ADMIN'] },
-        { section: 'budget', isEnabled: true },
-      ];
-
-      const sectionsIfAllowed = ['contribute', 'updates', 'budget'];
-      const sectionsIfNotAllowed = ['contribute', 'budget'];
-      expect(filterSectionsForUser(sections, false, false)).toEqual(sectionsIfNotAllowed);
-      expect(filterSectionsForUser(sections, true, false)).toEqual(sectionsIfAllowed);
-      expect(filterSectionsForUser(sections, false, true)).toEqual(sectionsIfAllowed);
-      expect(filterSectionsForUser(sections, true, true)).toEqual(sectionsIfAllowed);
-    });
-  });
-
-  describe('filterSectionsByData', () => {
-    describe('Contribute', () => {
-      it('Gets removed if not active (unless admin)', () => {
-        expect(filterSectionsByData(['contribute'], { isActive: true, features })).toEqual(['contribute']);
-        expect(filterSectionsByData(['contribute'], { isActive: false, features })).toEqual([]);
-        expect(filterSectionsByData(['contribute'], { isActive: false, features }, true)).toEqual(['contribute']);
-      });
-
-      it('Gets removed if no way to contribute (unless admin)', () => {
-        const collectiveWithoutContribs = { isActive: true, settings: { disableCustomContributions: true }, features };
-        expect(filterSectionsByData(['contribute'], collectiveWithoutContribs)).toEqual([]);
-        expect(filterSectionsByData(['contribute'], collectiveWithoutContribs, true)).toEqual(['contribute']);
-      });
-
-      it('Gets removed if event is passed (even if admin)', () => {
-        const event = { type: 'EVENT', isActive: true, features };
-        const pastDate = '2000-01-01';
-        const futureDate = new Date(new Date().getFullYear() + 1, 1).toUTCString();
-        expect(filterSectionsByData(['contribute'], { ...event, endsAt: pastDate })).toEqual([]);
-        expect(filterSectionsByData(['contribute'], { ...event, endsAt: pastDate }, true)).toEqual([]);
-        expect(filterSectionsByData(['contribute'], { ...event, endsAt: futureDate })).toEqual(['contribute']);
-      });
+      expect(addDefaultSections()).toEqual([]);
     });
   });
 
   describe('getFilteredSectionsForCollective', () => {
-    describe('For USER collective', () => {
+    describe('For event', () => {
+      const features = {
+        RECEIVE_FINANCIAL_CONTRIBUTIONS: 'ACTIVE',
+      };
+
+      it('Contribute gets removed if event is passed (even if admin)', () => {
+        const event = { type: 'EVENT', isActive: true, features };
+        const pastDate = '2000-01-01';
+        const futureDate = new Date(new Date().getFullYear() + 1, 1).toUTCString();
+        expectSectionsToContain(getFilteredSectionsForCollective({ ...event, endsAt: pastDate }), []);
+        expectSectionsToContain(getFilteredSectionsForCollective({ ...event, endsAt: pastDate }, true), []);
+        expectSectionsToContain(getFilteredSectionsForCollective({ ...event, endsAt: futureDate }), ['contribute']);
+      });
+    });
+    describe('For collective', () => {
+      const features = {
+        COLLECTIVE_GOALS: 'DISABLED',
+        CONVERSATIONS: 'DISABLED',
+        UPDATES: 'AVAILABLE',
+        RECEIVE_FINANCIAL_CONTRIBUTIONS: 'ACTIVE',
+        TRANSACTIONS: 'ACTIVE',
+      };
+
       it('Filters for non-admins', () => {
         const collective = { type: 'COLLECTIVE', features };
-        expect(getFilteredSectionsForCollective(collective)).toEqual(['contributors']);
-        expect(getFilteredSectionsForCollective({ ...collective, isActive: true })).toContain('contribute');
-        expect(getFilteredSectionsForCollective({ ...collective, longDescription: 'Hi' })).toContain('about');
+        expectSectionsToContain(getFilteredSectionsForCollective({ ...collective, isActive: true }), ['contribute']);
+        expectSectionsToContain(getFilteredSectionsForCollective({ ...collective, longDescription: 'Hi' }), ['about']);
       });
 
       it('Filters for admins', () => {
         const collective = { type: 'COLLECTIVE', features };
-        expect(getFilteredSectionsForCollective(collective, true)).toEqual([
+        expectSectionsToMatch(getFilteredSectionsForCollective(collective, true), [
           'contribute',
+          'contributors',
+          'top-financial-contributors',
           'updates',
           'budget',
           'contributors',
           'about',
+          'our-team',
         ]);
       });
-    });
-  });
-
-  describe('convertSectionsToNewFormat', () => {
-    it('converts to new format and preserves (as much as possible) the order', () => {
-      expect(
-        convertSectionsToNewFormat([
-          {
-            section: 'goals',
-            isEnabled: false,
-          },
-          {
-            section: 'contribute',
-            isEnabled: true,
-          },
-          {
-            section: 'updates',
-            isEnabled: false,
-            restrictedTo: [],
-          },
-          {
-            section: 'conversations',
-            isEnabled: false,
-            restrictedTo: [],
-          },
-          {
-            section: 'budget',
-            isEnabled: true,
-          },
-          {
-            section: 'contributors',
-            isEnabled: false,
-            restrictedTo: [],
-          },
-          {
-            section: 'recurring-contributions',
-            isEnabled: false,
-            restrictedTo: [],
-          },
-          {
-            section: 'about',
-            isEnabled: false,
-            restrictedTo: [],
-          },
-        ]),
-      ).toEqual([
-        {
-          type: 'SECTION',
-          isEnabled: false,
-          name: 'goals',
-          restrictedTo: null,
-        },
-        {
-          type: 'CATEGORY',
-          name: 'CONTRIBUTE',
-          isEnabled: true,
-          sections: [
-            {
-              isEnabled: true,
-              name: 'contribute',
-              restrictedTo: null,
-              type: 'SECTION',
-            },
-            {
-              isEnabled: true,
-              name: 'top-financial-contributors',
-              type: 'SECTION',
-            },
-            {
-              isEnabled: false,
-              name: 'contributors',
-              restrictedTo: [],
-              type: 'SECTION',
-            },
-          ],
-        },
-        {
-          type: 'CATEGORY',
-          name: 'CONNECT',
-          isEnabled: true,
-          sections: [
-            {
-              isEnabled: false,
-              name: 'updates',
-              restrictedTo: [],
-              type: 'SECTION',
-            },
-            {
-              isEnabled: false,
-              name: 'conversations',
-              restrictedTo: [],
-              type: 'SECTION',
-            },
-          ],
-        },
-        {
-          type: 'CATEGORY',
-          name: 'BUDGET',
-          isEnabled: true,
-          sections: [
-            {
-              isEnabled: true,
-              name: 'budget',
-              restrictedTo: null,
-              type: 'SECTION',
-            },
-          ],
-        },
-        {
-          type: 'CATEGORY',
-          name: 'CONTRIBUTIONS',
-          isEnabled: true,
-          sections: [
-            {
-              isEnabled: false,
-              name: 'recurring-contributions',
-              restrictedTo: [],
-              type: 'SECTION',
-            },
-          ],
-        },
-        {
-          type: 'CATEGORY',
-          name: 'ABOUT',
-          isEnabled: true,
-          sections: [
-            {
-              isEnabled: false,
-              name: 'about',
-              restrictedTo: [],
-              type: 'SECTION',
-            },
-          ],
-        },
-      ]);
     });
   });
 });
