@@ -5,41 +5,39 @@ import { Envelope } from '@styled-icons/boxicons-regular/Envelope';
 import { Planet } from '@styled-icons/boxicons-regular/Planet';
 import { Receipt } from '@styled-icons/boxicons-regular/Receipt';
 import { MoneyCheckAlt } from '@styled-icons/fa-solid/MoneyCheckAlt';
-import { Settings } from '@styled-icons/feather/Settings';
+import { AttachMoney } from '@styled-icons/material/AttachMoney';
 import { Close } from '@styled-icons/material/Close';
 import { Dashboard } from '@styled-icons/material/Dashboard';
 import { Stack } from '@styled-icons/remix-line/Stack';
 import themeGet from '@styled-system/theme-get';
 import { get, pickBy, without } from 'lodash';
-import { useRouter } from 'next/router';
 import { FormattedMessage, useIntl } from 'react-intl';
 import styled, { css } from 'styled-components';
 import { maxWidth } from 'styled-system';
 
-import { getFilteredSectionsForCollective, hasNewNavbar, NAVBAR_CATEGORIES } from '../../lib/collective-sections';
+import { getFilteredSectionsForCollective, NAVBAR_CATEGORIES } from '../../lib/collective-sections';
 import { CollectiveType } from '../../lib/constants/collectives';
 import useGlobalBlur from '../../lib/hooks/useGlobalBlur';
-import i18nCollectivePageSection from '../../lib/i18n-collective-page-section';
 
+import AddFundsBtn from '../AddFundsBtn';
+import AddPrepaidBudgetBtn from '../AddPrepaidBudgetBtn';
 import ApplyToHostBtn from '../ApplyToHostBtn';
 import Avatar from '../Avatar';
-import { AllSectionsNames, Dimensions } from '../collective-page/_constants';
-import CollectiveCallsToAction from '../CollectiveCallsToAction';
+import { Dimensions } from '../collective-page/_constants';
 import Container from '../Container';
 import { Box, Flex } from '../Grid';
 import Link from '../Link';
 import LinkCollective from '../LinkCollective';
 import LoadingPlaceholder from '../LoadingPlaceholder';
 import StyledButton from '../StyledButton';
-import StyledRoundButton from '../StyledRoundButton';
 import { H1, Span } from '../Text';
+import { useUser } from '../UserProvider';
 
 import CollectiveNavbarActionsMenu, { getContributeRoute } from './ActionsMenu';
 import { getNavBarMenu, NAVBAR_ACTION_TYPE } from './menu';
 import NavBarCategoryDropdown from './NavBarCategoryDropdown';
 
-// Nav v2 styled components
-const MainContainerV2 = styled(Container)`
+const MainContainer = styled(Container)`
   background: white;
   display: flex;
   justify-content: flex-start;
@@ -62,7 +60,7 @@ const AvatarBox = styled(Box)`
   }
 `;
 
-const InfosContainerV2 = styled(Container)`
+const InfosContainer = styled(Container)`
   [data-hide='false'] {
     width: 1;
     opacity: 1;
@@ -79,7 +77,7 @@ const InfosContainerV2 = styled(Container)`
   }
 `;
 
-const CollectiveNameV2 = styled(H1)`
+const CollectiveName = styled(H1)`
   letter-spacing: -0.8px;
 
   a {
@@ -120,97 +118,6 @@ const CategoriesContainer = styled(Container)`
   }
 `;
 
-// v1 components
-/** Main container for the entire component */
-const MainContainer = styled.div`
-  background: white;
-  ${props =>
-    props.withShadow &&
-    css`
-      box-shadow: 0px 6px 10px -5px rgba(214, 214, 214, 0.5);
-    `}
-
-  /** Everything's inside cannot be larger than max section width */
-  & > * {
-    max-width: ${Dimensions.MAX_SECTION_WIDTH}px;
-    margin: 0 auto;
-  }
-`;
-
-/** A single menu link */
-const MenuLink = styled.a`
-  display: block;
-  color: ${themeGet('colors.black.700')};
-  font-size: 14px;
-  line-height: 24px;
-  text-decoration: none;
-  white-space: nowrap;
-  padding: 12px 16px 16px;
-
-  letter-spacing: 0.6px;
-  text-transform: uppercase;
-  font-weight: 500;
-
-  &:focus {
-    color: ${themeGet('colors.primary.700')};
-    text-decoration: none;
-  }
-
-  &:hover {
-    color: ${themeGet('colors.primary.400')};
-    text-decoration: none;
-  }
-
-  @media (max-width: 64em) {
-    padding: 16px;
-  }
-`;
-
-const MenuLinkContainer = styled.div`
-  cursor: pointer;
-
-  &::after {
-    content: '';
-    display: block;
-    width: 0;
-    height: 3px;
-    background: ${themeGet('colors.primary.500')};
-    transition: width 0.2s;
-    float: right;
-  }
-
-  ${props =>
-    props.isSelected &&
-    css`
-      color: #090a0a;
-      font-weight: 500;
-      ${MenuLink} {
-        color: #090a0a;
-      }
-      @media (min-width: 64em) {
-        &::after {
-          width: 100%;
-          float: left;
-        }
-      }
-    `}
-
-  ${props =>
-    props.mobileOnly &&
-    css`
-      @media (min-width: 64em) {
-        display: none;
-      }
-    `}
-
-  @media (max-width: 64em) {
-    border-top: 1px solid #e1e1e1;
-    &::after {
-      display: none;
-    }
-  }
-`;
-
 /** Displayed on mobile & tablet to toggle the menu */
 const ExpandMenuIcon = styled(DotsVerticalRounded).attrs({ size: 28 })`
   cursor: pointer;
@@ -232,36 +139,6 @@ const ExpandMenuIcon = styled(DotsVerticalRounded).attrs({ size: 28 })`
   @media (min-width: 64em) {
     display: none;
   }
-`;
-
-const InfosContainer = styled(Container)`
-  padding: 14px 30px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  opacity: 1;
-  visibility: visible;
-  transform: translateY(0);
-  transition: opacity 0.075s ease-out, transform 0.1s ease-out, visibility 0.075s ease-out;
-
-  @media (max-width: 64em) {
-    padding: 10px 16px;
-  }
-
-  @media (min-width: 52em) {
-    ${ExpandMenuIcon} {
-      display: none;
-    }
-  }
-
-  /** Hidden state */
-  ${props =>
-    props.isHidden &&
-    css`
-      visibility: hidden;
-      opacity: 0;
-      transform: translateY(-20px);
-    `}
 `;
 
 const CloseMenuIcon = styled(Close).attrs({ size: 28 })`
@@ -286,62 +163,36 @@ const CloseMenuIcon = styled(Close).attrs({ size: 28 })`
   }
 `;
 
-const CollectiveName = styled.h1`
-  margin: 0 8px;
-  padding: 8px 0;
-  font-size: 20px;
-  line-height: 24px;
-  text-align: center;
-  letter-spacing: -1px;
-  font-weight: bold;
-  max-width: 50%;
-
-  a:not(:hover) {
-    color: #313233;
-  }
-
-  @media (min-width: 64em) {
-    text-align: center;
-  }
-`;
-
 const isFeatureAvailable = (collective, feature) => {
   const status = get(collective.features, feature);
   return status === 'ACTIVE' || status === 'AVAILABLE';
 };
 
-const getDefaultCallsToActions = (collective, isAdmin, newNavbarFeatureFlag) => {
+const getDefaultCallsToActions = (collective, isAdmin, isHostAdmin, isRoot) => {
   if (!collective) {
     return {};
   }
 
-  const isCollective = collective.type === CollectiveType.COLLECTIVE;
-  const isEvent = collective.type === CollectiveType.EVENT;
   const isFund = collective.type === CollectiveType.FUND;
-
-  if (newNavbarFeatureFlag) {
-    return {
-      hasContact: isFeatureAvailable(collective, 'CONTACT_FORM'),
-      hasApply: isFeatureAvailable(collective, 'RECEIVE_HOST_APPLICATIONS'),
-      hasSubmitExpense: isFeatureAvailable(collective, 'RECEIVE_EXPENSES'),
-      hasManageSubscriptions: isAdmin && get(collective.features, 'RECURRING_CONTRIBUTIONS') === 'ACTIVE',
-      hasDashboard: isAdmin && isFeatureAvailable(collective, 'HOST_DASHBOARD'),
-      hasRequestGrant: isFund || get(collective.settings, 'fundingRequest') === true,
-    };
-  }
+  const isProject = collective.type === CollectiveType.PROJECT;
 
   return {
-    hasContact: collective.canContact,
-    hasApply: collective.canApply && !isAdmin,
-    hasManageSubscriptions: isAdmin && !isCollective && !isEvent,
+    hasContribute: (isFund || isProject) && collective.isActive,
+    hasContact: isFeatureAvailable(collective, 'CONTACT_FORM'),
+    hasApply: isFeatureAvailable(collective, 'RECEIVE_HOST_APPLICATIONS'),
+    hasSubmitExpense: isFeatureAvailable(collective, 'RECEIVE_EXPENSES'),
+    hasManageSubscriptions: isAdmin && get(collective.features, 'RECURRING_CONTRIBUTIONS') === 'ACTIVE',
+    hasDashboard: isAdmin && isFeatureAvailable(collective, 'HOST_DASHBOARD'),
+    hasRequestGrant: isFund || get(collective.settings, 'fundingRequest') === true,
+    addPrepaidBudget: isRoot && collective.type === CollectiveType.ORGANIZATION,
+    addFunds: isHostAdmin,
   };
 };
 
 /**
  * Returns the main CTA that should be displayed as a button outside of the action menu in this component.
- * Returns the second CTA that should be displayed as a button in ActionsMenu.js if only 2 CTAs.
  */
-const getMainAction = (collective, isAdmin, callsToAction) => {
+const getMainAction = (collective, callsToAction) => {
   if (!collective || !callsToAction) {
     return null;
   }
@@ -361,7 +212,7 @@ const getMainAction = (collective, isAdmin, callsToAction) => {
         </Link>
       ),
     };
-  } else if (!isAdmin && callsToAction.includes('hasContribute') && getContributeRoute(collective)) {
+  } else if (callsToAction.includes('hasContribute') && getContributeRoute(collective)) {
     return {
       type: NAVBAR_ACTION_TYPE.CONTRIBUTE,
       component: (
@@ -375,64 +226,7 @@ const getMainAction = (collective, isAdmin, callsToAction) => {
         </Link>
       ),
     };
-  } else if (!isAdmin && callsToAction.includes('hasApply')) {
-    const plan = collective.plan || {};
-    return {
-      type: NAVBAR_ACTION_TYPE.APPLY,
-      component: (
-        <ApplyToHostBtn
-          hostSlug={collective.slug}
-          buttonRenderer={props => <MainActionBtn {...props} />}
-          hostWithinLimit={!plan.hostedCollectivesLimit || plan.hostedCollectives < plan.hostedCollectivesLimit}
-        />
-      ),
-    };
-  } else if (callsToAction.includes('hasSubmitExpense')) {
-    return {
-      type: NAVBAR_ACTION_TYPE.SUBMIT_EXPENSE,
-      component: (
-        <Link route="create-expense" params={{ collectiveSlug: collective.slug }}>
-          <MainActionBtn tabIndex="-1">
-            <Receipt size="1em" />
-            <Span ml={2}>
-              <FormattedMessage id="ExpenseForm.Submit" defaultMessage="Submit expense" />
-            </Span>
-          </MainActionBtn>
-        </Link>
-      ),
-    };
-  } else if (callsToAction.includes('hasManageSubscriptions')) {
-    return {
-      type: NAVBAR_ACTION_TYPE.MANAGE_SUBSCRIPTIONS,
-      component: (
-        <Link route="recurring-contributions" params={{ slug: collective.slug }}>
-          <MainActionBtn tabIndex="-1">
-            <Stack size="1em" />
-            <Span ml={2}>
-              <FormattedMessage id="menu.subscriptions" defaultMessage="Manage Contributions" />
-            </Span>
-          </MainActionBtn>
-        </Link>
-      ),
-    };
-  } else if (!isAdmin && callsToAction.includes('hasContact')) {
-    return {
-      type: NAVBAR_ACTION_TYPE.CONTACT,
-      component: (
-        <Link route="collective-contact" params={{ collectiveSlug: collective.slug }}>
-          <MainActionBtn tabIndex="-1">
-            <Envelope size="1em" />
-            <Span ml={2}>
-              <FormattedMessage id="Contact" defaultMessage="Contact" />
-            </Span>
-          </MainActionBtn>
-        </Link>
-      ),
-    };
-  }
-  // CTAs after this line are called when figuring out the second CTA if there are only 2.
-  // TO DO: Need to add addFunds and addPrepaidBudget CTAs.
-  else if (callsToAction.includes('hasApply')) {
+  } else if (callsToAction.includes('hasApply')) {
     const plan = collective.plan || {};
     return {
       type: NAVBAR_ACTION_TYPE.APPLY,
@@ -456,6 +250,80 @@ const getMainAction = (collective, isAdmin, callsToAction) => {
             </Span>
           </MainActionBtn>
         </Link>
+      ),
+    };
+  } else if (callsToAction.includes('hasSubmitExpense')) {
+    return {
+      type: NAVBAR_ACTION_TYPE.SUBMIT_EXPENSE,
+      component: (
+        <Link route="create-expense" params={{ collectiveSlug: collective.slug }}>
+          <MainActionBtn tabIndex="-1">
+            <Receipt size="1em" />
+            <Span ml={2}>
+              <FormattedMessage id="menu.submitExpense" defaultMessage="Submit Expense" />
+            </Span>
+          </MainActionBtn>
+        </Link>
+      ),
+    };
+  } else if (callsToAction.includes('hasManageSubscriptions')) {
+    return {
+      type: NAVBAR_ACTION_TYPE.MANAGE_SUBSCRIPTIONS,
+      component: (
+        <Link route="recurring-contributions" params={{ slug: collective.slug }}>
+          <MainActionBtn tabIndex="-1">
+            <Stack size="1em" />
+            <Span ml={2}>
+              <FormattedMessage id="menu.subscriptions" defaultMessage="Manage Contributions" />
+            </Span>
+          </MainActionBtn>
+        </Link>
+      ),
+    };
+  } else if (callsToAction.includes('hasContact')) {
+    return {
+      type: NAVBAR_ACTION_TYPE.CONTACT,
+      component: (
+        <Link route="collective-contact" params={{ collectiveSlug: collective.slug }}>
+          <MainActionBtn tabIndex="-1">
+            <Envelope size="1em" />
+            <Span ml={2}>
+              <FormattedMessage id="Contact" defaultMessage="Contact" />
+            </Span>
+          </MainActionBtn>
+        </Link>
+      ),
+    };
+  } else if (callsToAction.includes(NAVBAR_ACTION_TYPE.ADD_FUNDS) && collective.host) {
+    return {
+      type: NAVBAR_ACTION_TYPE.ADD_FUNDS,
+      component: (
+        <AddFundsBtn collective={collective} host={collective.host}>
+          {btnProps => (
+            <MainActionBtn {...btnProps}>
+              <AttachMoney size="1em" />
+              <Span>
+                <FormattedMessage id="menu.addFunds" defaultMessage="Add Funds" />
+              </Span>
+            </MainActionBtn>
+          )}
+        </AddFundsBtn>
+      ),
+    };
+  } else if (callsToAction.includes(NAVBAR_ACTION_TYPE.ADD_PREPAID_BUDGET)) {
+    return {
+      type: NAVBAR_ACTION_TYPE.ADD_PREPAID_BUDGET,
+      component: (
+        <AddPrepaidBudgetBtn collective={collective}>
+          {btnProps => (
+            <MainActionBtn {...btnProps}>
+              <AttachMoney size="1em" />
+              <Span>
+                <FormattedMessage id="menu.addPrepaidBudget" defaultMessage="Add Prepaid Budget" />
+              </Span>
+            </MainActionBtn>
+          )}
+        </AddPrepaidBudgetBtn>
       ),
     };
   } else {
@@ -503,14 +371,10 @@ const CollectiveNavbar = ({
   collective,
   isAdmin,
   isLoading,
-  showEdit,
   sections,
-  selected,
   selectedCategory,
-  LinkComponent,
   callsToAction,
   onCollectiveClick,
-  onSectionClick,
   hideInfosOnDesktop,
   onlyInfos,
   isAnimated,
@@ -518,15 +382,17 @@ const CollectiveNavbar = ({
   withShadow,
   useAnchorsForCategories,
 }) => {
-  const router = useRouter();
-  const newNavbarFeatureFlag = hasNewNavbar(get(router, 'query.navbarVersion'));
   const intl = useIntl();
   const [isExpanded, setExpanded] = React.useState(false);
-  sections = sections || getFilteredSectionsForCollective(collective, isAdmin, null, newNavbarFeatureFlag);
-  callsToAction = { ...getDefaultCallsToActions(collective, isAdmin, newNavbarFeatureFlag), ...callsToAction };
+  const { LoggedInUser } = useUser();
+  isAdmin = isAdmin || LoggedInUser?.canEditCollective(collective);
+  const isHostAdmin = LoggedInUser?.isHostAdmin(collective);
+  sections = sections || getFilteredSectionsForCollective(collective, isAdmin, isHostAdmin);
+  const isRoot = LoggedInUser?.isRoot;
+  callsToAction = { ...getDefaultCallsToActions(collective, isAdmin, isHostAdmin, isRoot), ...callsToAction };
   const actionsArray = Object.keys(pickBy(callsToAction, Boolean));
-  const mainAction = getMainAction(collective, isAdmin, actionsArray);
-  const secondAction = getMainAction(collective, isAdmin, without(actionsArray, mainAction?.type));
+  const mainAction = getMainAction(collective, actionsArray);
+  const secondAction = actionsArray.length === 2 && getMainAction(collective, without(actionsArray, mainAction?.type));
   const navbarRef = useRef();
 
   useGlobalBlur(navbarRef, outside => {
@@ -537,9 +403,8 @@ const CollectiveNavbar = ({
     }
   });
 
-  return newNavbarFeatureFlag ? (
-    // v2
-    <MainContainerV2
+  return (
+    <MainContainer
       flexDirection={['column', 'row']}
       px={[0, Dimensions.PADDING_X[1]]}
       mx="auto"
@@ -549,7 +414,7 @@ const CollectiveNavbar = ({
       maxHeight="100vh"
     >
       {/** Collective info */}
-      <InfosContainerV2 isAnimated={isAnimated} mr={[0, 2]} display="flex" alignItems="center" px={[3, 0]} py={[2, 1]}>
+      <InfosContainer isAnimated={isAnimated} mr={[0, 2]} display="flex" alignItems="center" px={[3, 0]} py={[2, 1]}>
         <Flex alignItems="center" data-hide={hideInfosOnDesktop}>
           {showBackButton && (
             <Box display={['none', 'block']} mr={2}>
@@ -570,7 +435,7 @@ const CollectiveNavbar = ({
             </LinkCollective>
           </AvatarBox>
           <Box display={['block', null, null, onlyInfos ? 'block' : 'none']}>
-            <CollectiveNameV2
+            <CollectiveName
               mx={2}
               py={2}
               fontSize={['16px', '20px']}
@@ -585,7 +450,7 @@ const CollectiveNavbar = ({
               ) : (
                 <LinkCollective collective={collective} onClick={onCollectiveClick} />
               )}
-            </CollectiveNameV2>
+            </CollectiveName>
           </Box>
         </Flex>
         {!onlyInfos && (
@@ -597,7 +462,7 @@ const CollectiveNavbar = ({
             )}
           </Box>
         )}
-      </InfosContainerV2>
+      </InfosContainer>
       {/** Main navbar items */}
 
       {!onlyInfos && (
@@ -630,7 +495,7 @@ const CollectiveNavbar = ({
             )}
           </CategoriesContainer>
 
-          {/* CTAs for v2 navbar */}
+          {/* CTAs */}
           <Container
             display={isExpanded ? 'flex' : ['none', 'flex']}
             flexDirection={['column', 'row']}
@@ -642,6 +507,7 @@ const CollectiveNavbar = ({
             {mainAction && (
               <Container display={['none', 'flex']} alignItems="center">
                 {mainAction.component}
+                {secondAction?.component || null}
               </Container>
             )}
             {!isLoading && (
@@ -649,7 +515,6 @@ const CollectiveNavbar = ({
                 collective={collective}
                 callsToAction={callsToAction}
                 hiddenActionForNonMobile={mainAction?.type}
-                secondAction={secondAction}
               />
             )}
             {!onlyInfos && (
@@ -663,117 +528,6 @@ const CollectiveNavbar = ({
             )}
           </Container>
         </Fragment>
-      )}
-    </MainContainerV2>
-  ) : (
-    // v1
-    <MainContainer withShadow={withShadow}>
-      {/** Collective infos */}
-      <InfosContainer isHidden={hideInfosOnDesktop} isAnimated={isAnimated}>
-        <Flex alignItems="center" flex="1 1 80%" css={{ minWidth: 0 /** For text-overflow */ }}>
-          <LinkCollective collective={collective} onClick={onCollectiveClick}>
-            <Container borderRadius="25%" mr={2}>
-              <Avatar collective={collective} radius={40} />
-            </Container>
-          </LinkCollective>
-          <CollectiveName>
-            {isLoading ? (
-              <LoadingPlaceholder height={22} minWidth={100} />
-            ) : (
-              <LinkCollective collective={collective} onClick={onCollectiveClick} />
-            )}
-          </CollectiveName>
-          {isAdmin && showEdit && (
-            <Link route="editCollective" params={{ slug: collective.slug }} title="Settings">
-              <StyledRoundButton size={24} bg="#F0F2F5" color="#4B4E52">
-                <Settings size={17} />
-              </StyledRoundButton>
-            </Link>
-          )}
-        </Flex>
-        {!onlyInfos && <ExpandMenuIcon onClick={() => setExpanded(!isExpanded)} />}
-      </InfosContainer>
-
-      {/** Navbar items and buttons */}
-      {!onlyInfos && (
-        <Container
-          position={['absolute', 'relative']}
-          display="flex"
-          justifyContent="space-between"
-          px={[0, Dimensions.PADDING_X[1]]}
-          width="100%"
-          background="white"
-        >
-          {isLoading ? (
-            <LoadingPlaceholder height={43} minWidth={150} mb={2} />
-          ) : (
-            <Container
-              flex="2 1 600px"
-              css={{ overflowX: 'auto' }}
-              display={isExpanded ? 'flex' : ['none', null, 'flex']}
-              data-cy="CollectivePage.NavBar"
-              flexDirection={['column', null, 'row']}
-              height="100%"
-              borderBottom={['1px solid #e6e8eb', 'none']}
-              backgroundColor="#fff"
-              zIndex={1}
-            >
-              {sections.map(section => (
-                <MenuLinkContainer
-                  key={section}
-                  isSelected={section === selected}
-                  onClick={() => {
-                    if (isExpanded) {
-                      setExpanded(false);
-                    }
-                    if (onSectionClick) {
-                      onSectionClick(section);
-                    }
-                  }}
-                >
-                  <MenuLink
-                    as={LinkComponent}
-                    collectivePath={collective.path || `/${collective.slug}`}
-                    section={section}
-                    label={i18nCollectivePageSection(intl, section)}
-                  />
-                </MenuLinkContainer>
-              ))}
-              {/* mobile CTAs */}
-              {callsToAction.hasSubmitExpense && (
-                <MenuLinkContainer mobileOnly>
-                  <MenuLink as={Link} route="create-expense" params={{ collectiveSlug: collective.slug }}>
-                    <FormattedMessage id="menu.submitExpense" defaultMessage="Submit Expense" />
-                  </MenuLink>
-                </MenuLinkContainer>
-              )}
-              {callsToAction.hasContact && (
-                <MenuLinkContainer mobileOnly>
-                  <MenuLink as={Link} route="collective-contact" params={{ collectiveSlug: collective.slug }}>
-                    <FormattedMessage id="Contact" defaultMessage="Contact" />
-                  </MenuLink>
-                </MenuLinkContainer>
-              )}
-              {callsToAction.hasDashboard && (
-                <MenuLinkContainer mobileOnly>
-                  <MenuLink as={Link} route="host.dashboard" params={{ hostCollectiveSlug: collective.slug }}>
-                    <FormattedMessage id="host.dashboard" defaultMessage="Dashboard" />
-                  </MenuLink>
-                </MenuLinkContainer>
-              )}
-            </Container>
-          )}
-          <div>
-            {!isLoading && (
-              // non-mobile CTAs
-              <CollectiveCallsToAction
-                display={['none', null, 'flex']}
-                collective={collective}
-                callsToAction={callsToAction}
-              />
-            )}
-          </div>
-        </Container>
       )}
     </MainContainer>
   );
@@ -805,12 +559,6 @@ CollectiveNavbar.propTypes = {
   isAdmin: PropTypes.bool,
   /** Will show loading state */
   isLoading: PropTypes.bool,
-  /** Wether we want to display the "/edit" button */
-  showEdit: PropTypes.bool,
-  /** Called with the new section name when it changes */
-  onSectionClick: PropTypes.func,
-  /** An optionnal function to build links URLs. Useful to override behaviour in test/styleguide envs. */
-  LinkComponent: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
   /** The list of sections to be displayed by the NavBar. If not provided, will show all the sections available to this collective type. */
   sections: PropTypes.arrayOf(
     PropTypes.shape({
@@ -820,8 +568,7 @@ CollectiveNavbar.propTypes = {
   ),
   /** Called when users click the collective logo or name */
   onCollectiveClick: PropTypes.func,
-  /** Currently selected section */
-  selected: PropTypes.oneOf(AllSectionsNames),
+  /** Currently selected category */
   selectedCategory: PropTypes.oneOf(Object.values(NAVBAR_CATEGORIES)),
   /** If true, the collective infos (avatar + name) will be hidden with css `visibility` */
   hideInfosOnDesktop: PropTypes.bool,
@@ -846,15 +593,6 @@ CollectiveNavbar.defaultProps = {
   callsToAction: {},
   showBackButton: true,
   withShadow: true,
-
-  // eslint-disable-next-line react/prop-types
-  LinkComponent: function DefaultNavbarLink({ section, label, collectivePath, className }) {
-    return (
-      <Link route={`${collectivePath}#section-${section}`} className={className}>
-        {label}
-      </Link>
-    );
-  },
 };
 
 export default React.memo(CollectiveNavbar);
