@@ -45,21 +45,19 @@ class SectionProjects extends React.PureComponent {
   }
 
   render() {
-    const { collective, projects, isAdmin, showTitle } = this.props;
+    const { collective, projects, isAdmin } = this.props;
 
     if ((projects.length === 0 || !collective.isActive) && !isAdmin) {
       return null;
     }
 
     return (
-      <Box pt={showTitle ? [4, 5] : [2, 3]} data-cy="Projects">
-        {showTitle && (
-          <ContainerSectionContent>
-            <SectionTitle>
-              <FormattedMessage id="CollectivePage.SectionProjects.Title" defaultMessage="Projects" />
-            </SectionTitle>
-          </ContainerSectionContent>
-        )}
+      <Box pt={[4, 5]} data-cy="Projects">
+        <ContainerSectionContent>
+          <SectionTitle>
+            <FormattedMessage id="CollectivePage.SectionProjects.Title" defaultMessage="Projects" />
+          </SectionTitle>
+        </ContainerSectionContent>
 
         <Box mb={4}>
           <HorizontalScroller getScrollDistance={this.getContributeCardsScrollDistance}>
