@@ -109,7 +109,7 @@ FeesBreakdown.propTypes = {
   platformFeePercent: PropTypes.object,
   hostFeePercent: PropTypes.object,
   paymentMethod: PropTypes.object,
-  currency: PropTypes.object,
+  currency: PropTypes.string,
 };
 
 FeesBreakdown.defaultProps = {
@@ -149,7 +149,7 @@ const getTaxPerentageForProfile = (taxes, tierType, hostCountry, collectiveCount
   }
 };
 
-const VATInputs = ({ collective, taxInfo, dispatchChange, setFormState, formState }) => {
+const VATInputs = ({ currency, taxInfo, dispatchChange, setFormState, formState }) => {
   const hasConfirmedTaxID = taxInfo.number && taxInfo.isReady;
   const vatShortLabel = <FormattedMessage id="tax.vatShort" defaultMessage="VAT" />;
 
@@ -275,7 +275,7 @@ const VATInputs = ({ collective, taxInfo, dispatchChange, setFormState, formStat
           )}
         </Flex>
         <Span fontSize="16px" pt={2}>
-          {taxInfo.isReady && `+ ${formatCurrency(taxInfo.amount, collective.currency)}`}
+          {taxInfo.isReady && `+ ${formatCurrency(taxInfo.amount, currency)}`}
         </Span>
       </AmountLine>
     </React.Fragment>
@@ -285,7 +285,7 @@ const VATInputs = ({ collective, taxInfo, dispatchChange, setFormState, formStat
 VATInputs.propTypes = {
   formState: PropTypes.object,
   taxInfo: PropTypes.object,
-  collective: PropTypes.object,
+  currency: PropTypes.string,
   dispatchChange: PropTypes.func,
   setFormState: PropTypes.func,
 };
@@ -309,6 +309,8 @@ const StepSummary = ({
   const tierType = tier?.type;
   const hostCountry = get(collective.host, 'location.country');
   const collectiveCountry = collective.location.country || get(collective.parent, 'location.country');
+
+  const currency = tier?.amount.currency || collective.currency;
 
   const [formState, setFormState] = useState({ isEnabled: false, error: false });
   const taxPercentage = getTaxPerentageForProfile(taxes, tierType, hostCountry, collectiveCountry, data);
@@ -340,7 +342,7 @@ const StepSummary = ({
         <FeesBreakdown
           platformFeePercent={collective.platformFeePercent}
           hostFeePercent={collective.hostFeePercent}
-          currency={collective.currency}
+          currency={currency}
           paymentMethod={paymentMethod}
           amount={amount}
         />
@@ -358,11 +360,7 @@ const StepSummary = ({
               <FormattedMessage id="contribution.itemPrice" defaultMessage="Item price" />
             </Label>
             <Span fontSize="16px">
-              {amount ? (
-                formatCurrency(amount, collective.currency)
-              ) : (
-                <FormattedMessage id="Amount.Free" defaultMessage="Free" />
-              )}
+              {amount ? formatCurrency(amount, currency) : <FormattedMessage id="Amount.Free" defaultMessage="Free" />}
             </Span>
           </AmountLine>
         </React.Fragment>
@@ -371,13 +369,13 @@ const StepSummary = ({
         <Label fontWeight="bold">
           <FormattedMessage id="contribution.your" defaultMessage="Your contribution" />
         </Label>
-        <Span fontSize="16px">{formatCurrency(amount * quantity, collective.currency)}</Span>
+        <Span fontSize="16px">{formatCurrency(amount * quantity, currency)}</Span>
       </AmountLine>
       {applyTaxes && amount > 0 && (
         <div>
           {taxes.some(({ type }) => type === TaxType.VAT) && (
             <VATInputs
-              collective={collective}
+              currency={currency}
               dispatchChange={dispatchChange}
               setFormState={setFormState}
               formState={formState}
@@ -390,7 +388,7 @@ const StepSummary = ({
                 GST ({taxPercentage}%)
               </Span>
               <Span fontSize="16px" pt={2}>
-                {`+ ${formatCurrency(taxInfo.amount, collective.currency)}`}
+                {`+ ${formatCurrency(taxInfo.amount, currency)}`}
               </Span>
             </AmountLine>
           )}
@@ -402,7 +400,7 @@ const StepSummary = ({
           <FormattedMessage id="contribution.total" defaultMessage="TOTAL" />
         </Label>
         <Span fontWeight="bold" fontSize="16px" color={taxInfo.isReady ? 'black.800' : 'black.400'}>
-          {formatCurrency(amount * quantity + (taxInfo.isReady ? taxInfo.amount : 0), collective.currency)}
+          {formatCurrency(amount * quantity + (taxInfo.isReady ? taxInfo.amount : 0), currency)}
         </Span>
       </AmountLine>
     </Box>
@@ -422,7 +420,6 @@ StepSummary.propTypes = {
     }),
   }),
   collective: PropTypes.shape({
-    /** The currency used for the transaction */
     currency: PropTypes.string.isRequired,
     /** Host fees, as an integer percentage */
     hostFeePercent: PropTypes.number,
@@ -458,6 +455,9 @@ StepSummary.propTypes = {
   /** Type of the tier. Used to check if taxes apply */
   tier: PropTypes.shape({
     type: PropTypes.oneOf(tiersTypes),
+    amount: PropTypes.shape({
+      currency: PropTypes.string,
+    }),
   }),
   /** Payment method, used to generate label and payment fee */
   paymentMethod: PropTypes.shape({
