@@ -110,9 +110,11 @@ const AddFundsModal = ({ host, collective, ...props }) => {
 
   // From the Collective page we pass host and collective as API v1 objects
   // From the Host dashboard we pass host and collective as API v2 objects
-  const canAddHostFee = host.plan.hostFees && collective.id !== host.id;
-
+  const canAddHostFee = host.plan?.hostFees && collective.id !== host.id;
   const defaultHostFeePercent = canAddHostFee ? collective.hostFeePercent : 0;
+
+  const canAddPlatformFee = LoggedInUser.isRoot();
+  const defaultPlatformFeePercent = 0;
 
   if (!LoggedInUser) {
     return null;
@@ -137,7 +139,9 @@ const AddFundsModal = ({ host, collective, ...props }) => {
       >
         {({ values, isSubmitting, isValid, dirty }) => {
           const hostFeePercent = isNaN(values.hostFeePercent) ? defaultHostFeePercent : values.hostFeePercent;
-          const platformFeePercent = isNaN(values.platformFeePercent) ? 0 : values.platformFeePercent;
+          const platformFeePercent = isNaN(values.platformFeePercent)
+            ? defaultPlatformFeePercent
+            : values.platformFeePercent;
           const hostFee = Math.round(values.amount * (hostFeePercent / 100));
           const platformFee = Math.round(values.amount * (platformFeePercent / 100));
 
@@ -161,7 +165,7 @@ const AddFundsModal = ({ host, collective, ...props }) => {
                     {({ form, field }) => (
                       <StyledInputAmount
                         id={field.id}
-                        currency={collective.currency || host.currency}
+                        currency={collective.currency}
                         placeholder="0.00"
                         error={field.error}
                         value={field.value}
@@ -191,7 +195,7 @@ const AddFundsModal = ({ host, collective, ...props }) => {
                     </StyledInputFormikField>
                   )}
                 </Flex>
-                {LoggedInUser.isRoot() && (
+                {canAddPlatformFee && (
                   <StyledInputFormikField
                     name="platformFeePercent"
                     htmlFor="addFunds-platformFeePercent"
@@ -243,35 +247,39 @@ const AddFundsModal = ({ host, collective, ...props }) => {
                 <StyledHr my={2} borderColor="black.300" />
                 <AmountDetailsLine
                   value={values.amount || 0}
-                  currency={host.currency}
-                  label={<FormattedMessage id="addfunds.totalAmount" defaultMessage="Amount" />}
+                  currency={collective.currency}
+                  label={<FormattedMessage id="addfunds.totalAmount" defaultMessage="Funding amount" />}
                 />
-                <AmountDetailsLine
-                  value={hostFee}
-                  currency={host.currency}
-                  label={
-                    <FormattedMessage
-                      id="addfunds.hostFees"
-                      defaultMessage="Host fees ({hostFees})"
-                      values={{ hostFees: `${hostFeePercent}%` }}
-                    />
-                  }
-                />
-                <AmountDetailsLine
-                  value={platformFee}
-                  currency={host.currency}
-                  label={
-                    <FormattedMessage
-                      id="addfunds.platformFees"
-                      defaultMessage="Platform fees ({platformFees})"
-                      values={{ platformFees: `${platformFeePercent}%` }}
-                    />
-                  }
-                />
+                {canAddHostFee && (
+                  <AmountDetailsLine
+                    value={hostFee}
+                    currency={collective.currency}
+                    label={
+                      <FormattedMessage
+                        id="addfunds.hostFees"
+                        defaultMessage="Host fees ({hostFees})"
+                        values={{ hostFees: `${hostFeePercent}%` }}
+                      />
+                    }
+                  />
+                )}
+                {canAddPlatformFee && (
+                  <AmountDetailsLine
+                    value={platformFee}
+                    currency={collective.currency}
+                    label={
+                      <FormattedMessage
+                        id="addfunds.platformFees"
+                        defaultMessage="Platform fees ({platformFees})"
+                        values={{ platformFees: `${platformFeePercent}%` }}
+                      />
+                    }
+                  />
+                )}
                 <StyledHr my={2} borderColor="black.300" />
                 <AmountDetailsLine
                   value={values.amount - hostFee - platformFee}
-                  currency={host.currency}
+                  currency={collective.currency}
                   label={<FormattedMessage id="addfunds.netAmount" defaultMessage="Net amount" />}
                   isLargeAmount
                 />
@@ -287,6 +295,7 @@ const AddFundsModal = ({ host, collective, ...props }) => {
               <ModalFooter isFullWidth>
                 <Flex justifyContent="center" flexWrap="wrap">
                   <StyledButton
+                    type="submit"
                     buttonStyle="primary"
                     mx={2}
                     mb={1}
@@ -296,7 +305,7 @@ const AddFundsModal = ({ host, collective, ...props }) => {
                   >
                     <FormattedMessage id="menu.addFunds" defaultMessage="Add Funds" />
                   </StyledButton>
-                  <StyledButton mx={2} mb={1} minWidth={100} onClick={props.onClose}>
+                  <StyledButton mx={2} mb={1} minWidth={100} onClick={props.onClose} type="button">
                     <FormattedMessage id="actions.cancel" defaultMessage="Cancel" />
                   </StyledButton>
                 </Flex>
@@ -312,19 +321,16 @@ const AddFundsModal = ({ host, collective, ...props }) => {
 AddFundsModal.propTypes = {
   host: PropTypes.shape({
     id: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
-    legacyId: PropTypes.number,
-    currency: PropTypes.string,
     name: PropTypes.string,
-    hostFeePercent: PropTypes.number,
     plan: PropTypes.shape({
       hostFees: PropTypes.bool,
     }),
   }).isRequired,
   collective: PropTypes.shape({
     id: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
-    legacyId: PropTypes.number,
     currency: PropTypes.string,
     hostFeePercent: PropTypes.number,
+    platformFeePercent: PropTypes.number,
     slug: PropTypes.string,
   }).isRequired,
   onClose: PropTypes.func,
