@@ -5,7 +5,6 @@ import { get } from 'lodash';
 import dynamic from 'next/dynamic';
 import { createGlobalStyle } from 'styled-components';
 
-import { hasNewNavbar } from '../lib/collective-sections';
 import { generateNotFoundError } from '../lib/errors';
 
 import CollectivePageContent from '../components/collective-page';
@@ -53,7 +52,7 @@ const GlobalStyles = createGlobalStyle`
  * to render `components/collective-page` with everything needed.
  */
 class CollectivePage extends React.Component {
-  static async getInitialProps({ client, req, res, query: { slug, status, step, mode, navbarVersion } }) {
+  static async getInitialProps({ client, req, res, query: { slug, status, step, mode } }) {
     if (res && req && (req.language || req.locale === 'en')) {
       res.set('Cache-Control', 'public, s-maxage=300');
     }
@@ -62,8 +61,7 @@ class CollectivePage extends React.Component {
 
     // If on server side
     if (req) {
-      const hasNewCollectiveNavbar = hasNewNavbar(navbarVersion);
-      await preloadCollectivePageGraphlQueries(slug, client, hasNewCollectiveNavbar);
+      await preloadCollectivePageGraphlQueries(slug, client);
       skipDataFromTree = true;
     }
 
