@@ -305,12 +305,9 @@ Cypress.Commands.add('checkStepsProgress', ({ enabled = [], disabled = [] }) => 
 /**
  * Check if user is logged in by searching for its username in navbar
  */
-Cypress.Commands.add('assertLoggedIn', (username, timeout) => {
+Cypress.Commands.add('assertLoggedIn', params => {
   cy.log('Ensure user is logged in');
-  const loginSection = cy.get('.LoginTopBarProfileButton-name', { timeout });
-  if (username) {
-    loginSection.should('contain', username);
-  }
+  cy.getByDataCy('topbar-login-username', params);
 });
 
 /**
